versão inicial dos totalizadores

+++ VaquinhaWebAPI/DTOs/DespesaViewModel.cs
@@ -5,7 +5,7 @@ namespace VaquinhaWebAPI.DTOs
         public    int     PagamentoId {get; set;}
         public DateTime DtItemDespesa { get; set; }
         public string DescricaoItemDespesa { get; set; }
-        public decimal ValorItemDespesa { get; set; }
+        public double ValorItemDespesa { get; set; }
         public int ItemDespesaId {get; set;}
         public int PaganteId { get; set; }
         public string NomePagante { get; set; }
